Refactor HomePage Class, Delete DropdownMenu Class
