+++ src/test/java/model/HomePage.java
@@ -10,6 +10,8 @@ import org.openqa.selenium.support.ui.ExpectedConditions;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import static runner.TestUtils.scrollToElement;
+
 public class HomePage extends BasePage {
 
 
@@ -39,19 +41,21 @@ public class HomePage extends BasePage {
 
     @FindBy(css = "a[href='/me/my-views']")
     private WebElement myViews;
-    
-      @FindBy(xpath = "//a[@href='/manage']")
+
+    @FindBy(xpath = "//a[@href='/manage']")
     private WebElement manageJenkins;
 
     @FindBy(xpath = "//span/a[@href='/asynchPeople/']")
     private WebElement people;
 
-
-    @FindBy(css=".tabBar>.tab>a[class='']")
+    @FindBy(css = ".tabBar>.tab>a[class='']")
     private WebElement openViewLink;
 
-    @FindBy(css=".tabBar>.tab>a.addTab")
+    @FindBy(css = ".tabBar>.tab>a.addTab")
     private WebElement addViewLink;
+
+    @FindBy(xpath = "//span[text()='Move']")
+    private WebElement moveButtonDropdown;
 
     public HomePage(WebDriver driver) {
         super(driver);
@@ -118,15 +122,15 @@ public class HomePage extends BasePage {
         return new PipelineConfigPage(getDriver());
     }
 
-    public String getTextHeader(){
+    public String getTextHeader() {
         return header.getText();
     }
 
-    public DropdownMenu clickFolderDropdownMenu(String folderName) {
+    public HomePage clickFolderDropdownMenu(String folderName) {
         getWait(5).until(ExpectedConditions
                 .elementToBeClickable(By.xpath("//a[@href='job/" + folderName + "/']/button"))).click();
 
-        return new DropdownMenu(getDriver());
+        return this;
     }
 
     public FolderStatusPage clickFolder(String folderName) {
@@ -140,27 +144,34 @@ public class HomePage extends BasePage {
 
         return new ManageJenkinsPage(getDriver());
     }
-    
-     public MyViewsPage clickMyViews() {
+
+    public MyViewsPage clickMyViews() {
         myViews.click();
 
         return new MyViewsPage(getDriver());
     }
 
-    public ManageJenkinsPage clickManageJenkins(){
+    public ManageJenkinsPage clickManageJenkins() {
         manageJenkins.click();
 
         return new ManageJenkinsPage(getDriver());
     }
 
-    public PeoplePage clickPeople(){
+    public PeoplePage clickPeople() {
         people.click();
 
         return new PeoplePage(getDriver());
     }
 
-    public MultiConfigurationProjectStatusPage clickMultConfJobName(String name){
+    public MultiConfigurationProjectStatusPage clickMultConfJobName(String name) {
         jobList.get(0).click();
         return new MultiConfigurationProjectStatusPage(getDriver());
+    }
+
+    public MovePage clickMoveButtonDropdown() {
+        getWait(5).until(ExpectedConditions.visibilityOf(moveButtonDropdown));
+        scrollToElement(getDriver(), moveButtonDropdown);
+        moveButtonDropdown.click();
+        return new MovePage(getDriver());
     }
 }
